move syncjobs to package
