+++ schema/site.schema.json
@@ -997,7 +997,7 @@
     "authz.syncJobsLogsTTL": {
       "description": "EXPERIMENTAL: Time interval (in minutes) of how long to keep sync job logs for.",
       "type": "integer",
-      "default": 30
+      "default": 15
     },
     "externalService.userMode": {
       "description": "Enable to allow users to add external services for public and private repositories to the Sourcegraph instance.",
